1. a small bug don't figure out how to solve it now:
After doing the animation, the node included in animation cannot change color, when you activate it or inactivate. This issue maybe relate to animation function. Because I don't know the what's the different before animation and after animation.

2. the connection active can work perfer unless you refresh the page. When you refresh the page,  the connection will refresh that status, for example,  before refesh the page, the connection status is inActiveted, which represent as red color, after you refresh the page, the connection status in data is also inActivated, but it represent as white color like activated status.

so before refresh the page, change the connection status to the original status when get the network page in order to keep consistent between the popup information and database